+++ src/main/java/org/intellij/lang/annotations/Pattern.java
@@ -16,10 +16,10 @@
 
 package org.intellij.lang.annotations;
 
+import static java.lang.annotation.ElementType.*;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import static java.lang.annotation.ElementType.*;
 
 @Retention(RetentionPolicy.CLASS)
 @Target({ METHOD, FIELD, PARAMETER, LOCAL_VARIABLE, ANNOTATION_TYPE })