+++ src/main/java/org/jetbrains/annotations/NotNull.java
@@ -16,14 +16,10 @@
 
 package org.jetbrains.annotations;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
- * An element annotated with NotNull claims <code>null</code> value is <em>forbidden</em>
+ * An element annotated with NotNull claims {@code null} value is <em>forbidden</em>
  * to return (for methods), pass to (parameters) and hold (local variables and fields).
  * Apart from documentation purposes this annotation is intended to be used by static analysis tools
  * to validate against probable runtime errors and element contract violations.