+++ src/main/java/org/jetbrains/annotations/Nullable.java
@@ -16,11 +16,7 @@
 
 package org.jetbrains.annotations;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * An element annotated with {@link Nullable} claims {@code null} value is perfectly <em>valid</em>