+++ src/main/java/org/jetbrains/annotations/PropertyKey.java
@@ -16,11 +16,7 @@
 
 package org.jetbrains.annotations;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Specifies that a method parameter accepts arguments which must be valid property