+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Remove deployment from TravisCI infrastructure as it is no longer feasible.
+* Import source from the `org.jetbrains:annotations:jar:17.0.0` artifact available in the upstream repository [JetBrains/java-annotations](https://github.com/JetBrains/java-annotations).
 
 ## 1.0.0: